expose, correct, unify defaults and wire up (missing) playfield/flipper physics im/export settings
git-svn-id: https://svn.code.sf.net/p/vpinball/code/trunk@2516 14438d79-0ffd-4333-8ae7-ea2bdd901387

+++ light.cpp
@@ -799,7 +799,7 @@ void Light::PrepareMoversCustom()
       ShowError(textBuffer);
       return;
    }
-   customMoverIndexNum = vtri.size();
+   customMoverIndexNum = (int)vtri.size();
 
    g_pplayer->m_pin3d.m_pd3dDevice->CreateIndexBuffer(customMoverIndexNum, 0, IndexBuffer::FMT_INDEX16, &customMoverIBuffer);
